requirements.txt
# Prophet Forecasting Pipeline Dependencies

# Core ML and Data Science
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0

# Prophet and Time Series
prophet>=1.1.0
holidays>=0.18

# SARIMAX and Time Series Analysis
# Fixing numpy version for pmdarima compatibility
numpy<2.0,>=1.21.0
pmdarima>=2.0.0
statsmodels>=0.13.0
scipy>=1.9.0

# Hyperparameter Optimization
optuna>=3.0.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Configuration and Utilities
pyyaml>=6.0
click>=8.0.0
tqdm>=4.64.0
joblib>=1.2.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Logging and Monitoring
loguru>=0.6.0

# Data Validation
pandera>=0.15.0

# Additional utilities
python-dateutil>=2.8.0
pytz>=2022.1
